20210823 Telegraf - old-menu branch - PR 2 of 3
1. Implements self-repair functionality by adding `Dockerfile` and
`entrypoint.sh` to the Telegraf template.
2. Removes `telegraf.conf` from Telegraf template. The default
configuration file is now the one supplied with the container
downloaded from DockerHub. In other words, it will always be kept
up-to-date by the container's maintainers. Also note the lack of a
`directoryfix.sh` mechanism, implying that it was always the user's
responsibility to copy `telegraf.conf` into place.
3. Alters service definition to:

	* use a persistent storage area in `./volumes/telegraf`
	* remove the old `./services/telegraf/telegraf.conf` file mapping
	* add explicit mappings for all exposed ports.