20210924 Portainer - old-menu branch - PR 2 of 3
In [Issue 289](https://github.com/SensorsIot/IOTstack/issues/289),
@wijnsema proposed removal of Portainer. There has been additional
support for this from @mats-nk. It is coming up to a year since
Portainer-CE was added to IOTstack so this PR is my way of saying that
I agree with the idea.

1. Remove Portainer service definition.
2. Remove Portainer documentation.
3. Reduce Portainer-CE documentation to a stub pointing to revised
documentation on master branch.
4. Remove `directoryfix.sh` (Portainer to Portainer-CE migration aid).
5. Change port number to 9000. Old-menu originally had 9002 (so that it
could coexist with Portainer on 9000) but both new menu and
experimentatal branch used 9000 for both. This simply harmonises old-menu.
6. Removes references to `portainer` from `menu.sh`.